amadeus_request/requirements.txt
Flask==2.1.2
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
google-cloud-secret-manager
amadeus
pandas
google-cloud-language==2.4.1
googletrans==3.1.0a0